Rename implementation header file.
